requirements.txt
streamlit
pandas
plotly
langchain
langchain-experimental
langchain-google-genai
google-api-python-client
openpyxl
matplotlib
tabulate
python-dotenv
pytest
pytest-mock
psycopg2-binary
google-cloud-bigquery
snowflake-connector-python
ydata-profiling
streamlit-elements